Initial commit for Additional Theming Variables
This addition is meant to add a layer of theming into Apollo without disrupting current Apollo implementation.

Theme variables are meant to be updated at the base values (50-1000) and the semantically named colors will update automatically.

This is also meant to work in conjunction with luminosity testing to identify colors for text on colors that provide sufficient contrast

See a demo for the colors in use here:
https://codepen.io/burnt-rocket/pen/bGvjJxW?editors=1100